# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: RISE API
  version: 1.0.0
  description: RISE — Autonomous Incident Remediation System API Specification
servers:
  - url: https://api.rise.internal/api/v1
    description: Internal API Gateway
security:
  - BearerAuth: []
paths:
...
  /incidents:
    get:
      summary: List incidents with filters
      operationId: listIncidents
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, investigating, awaiting_approval, resolved, closed]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [SEV1, SEV2, SEV3, SEV4]
        - name: service
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            default: 50
      responses:
        '200':
          description: Paginated list of incidents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'

    post:
      summary: Manually create an incident
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreateRequest'
      responses:
        '201':
          description: Incident created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}:
    get:
      summary: Full incident detail
      operationId: getIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Expanded incident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetailEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'

    patch:
      summary: Update severity or status manually
      operationId: patchIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentUpdateRequest'
      responses:
        '200':
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /incidents/{incident_id}/reinvestigate:
    post:
      summary: Re-trigger agent pipeline for an incident
      operationId: reinvestigateIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Agent pipeline re-investigation queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReinvestigateEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
        '429':
          $ref: '#/components/responses/429RateLimited'
  /incidents/{incident_id}/comment:
    post:
      summary: Add human comment to incident timeline
      operationId: addIncidentComment
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
  /integrations:
    get:
      summary: List integrations
      operationId: listIntegrations
      responses:
        '200':
          description: List of connected integrations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationListEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
  /integrations/{type}/connect:
    post:
      summary: Initiate OAuth/App connection flow
      operationId: connectIntegration
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth redirect URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConnectEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
  /integrations/{type}:
    delete:
      summary: Disconnect an integration
      operationId: deleteIntegration
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Integration disconnected
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '422':
          $ref: '#/components/responses/422ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Meta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        next_cursor:
          type: string
          nullable: true
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_POLICY_VIOLATION
            - APPROVAL_EXPIRED
            - ACTION_PLAN_CHANGED
            - RESOURCE_LOCKED
            - INTEGRATION_UNAVAILABLE
            - RATE_LIMITED
            - UNAUTHORIZED
            - NOT_FOUND
        message:
          type: string
        details:
          type: object
    ErrorEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
          default: null
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    Incident:
      type: object
      required: [id, title, description, severity, status, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [SEV1, SEV2, SEV3, SEV4]
        status:
          type: string
          enum: [open, investigating, awaiting_approval, resolved, closed]
        affected_service:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        resolution_note:
          type: string
          nullable: true
    IncidentDetail:
      allOf:
        - $ref: '#/components/schemas/Incident'
        - type: object
          properties:
            timeline:
              type: array
              items:
                type: object
            root_cause:
              type: object
              nullable: true
            impact:
              type: object
              nullable: true
            actions:
              type: array
              items:
                type: object
            approvals:
              type: array
              items:
                type: object
    IncidentCreateRequest:
      type: object
      required: [title, description, severity, affected_service]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        severity:
          type: string
          enum: [SEV1, SEV2, SEV3, SEV4]
        affected_service:
          type: string
    IncidentUpdateRequest:
      type: object
      required: [status, resolution_note]
      properties:
        status:
          type: string
        resolution_note:
          type: string
    CommentCreateRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
    Comment:
      type: object
      required: [id, incident_id, text, created_at, author]
      properties:
        id:
          type: string
        incident_id:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
        author:
          type: string
    CommentEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/Comment'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    IncidentListEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    IncidentEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/Incident'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    IncidentDetailEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/IncidentDetail'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    ReinvestigateData:
      type: object
      required: [agent_run_id, status]
      properties:
        agent_run_id:
          type: string
        status:
          type: string
    ReinvestigateEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/ReinvestigateData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
    Integration:
      type: object
      required: [type, status, scopes]
      properties:
        type:
          type: string
        status:
          type: string
        scopes:
          type: array
          items:
            type: string
    IntegrationListEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Integration'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
    IntegrationConnectData:
      type: object
      required: [redirect_url]
      properties:
        redirect_url:
          type: string
    IntegrationConnectEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          $ref: '#/components/schemas/IntegrationConnectData'
        meta:
          $ref: '#/components/schemas/Meta'
        error:
          type: object
          nullable: true
...
  responses:
    401Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "UNAUTHORIZED"
              message: "Missing or invalid authorization token"
              details: {}
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "NOT_FOUND"
              message: "Requested resource was not found"
              details: {}
...
    422ValidationError:
      description: Unprocessable Entity - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "VALIDATION_ERROR"
              message: "Request body failed schema validation"
              details: {}
    429RateLimited:
      description: Too Many Requests - rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            meta:
              request_id: "00000000-0000-0000-0000-000000000000"
              timestamp: "2026-08-01T00:00:00Z"
            error:
              code: "RATE_LIMITED"
              message: "Rate limit exceeded"
              details: {}
...